+++ src/constants/strings.js
@@ -1,0 +1,3 @@
+export const CSR_Post_STRINGS = {
+  myPost: "My Posts",
+};
