+++ src/pages/post/first-post.js
@@ -9,7 +9,6 @@ export default function FirstPost() {
         <title>First Post</title>
       </Head>
       <h1>First Post</h1>
-      <h2>{/* <Link href="/">← Back to home</Link> */}</h2>
     </Layout>
   );
 }
